Force and angle shows countdown

+++ Assets/Scripts/Gameplay/Controllers/ForceController.cs
@@ -6,10 +6,11 @@ namespace Gameplay.Controllers
 {
     public class ForceController : MonoBehaviour
     {
-        [Header("Settings")] [SerializeField] private float maxForce = 100f;
+        [Header("Settings")] 
+        [SerializeField] private float maxForce = 100f;
         [SerializeField, Range(0, 1)] private float incrementPercentage = 0.1f;
         [SerializeField, Range(0, 1)] private float decrementPercentage = 0.1f;
-        [SerializeField] private float forceChargeTime = 3;
+        [SerializeField] private int forceChargeTime = 3;
         [SerializeField] private AnimationCurve chargeCurve;
 
         [SerializeField] private Button clickBtn;
@@ -18,7 +19,7 @@ namespace Gameplay.Controllers
         public float MaxForce => maxForce;
         public float IncrementPercentage => incrementPercentage;
         public float DecrementPercentage => decrementPercentage;
-        public float ForceChargeTime => forceChargeTime;
+        public int ForceChargeTime => forceChargeTime;
 
         public AnimationCurve ChargeCurve => chargeCurve;
